Title: 나무박멸, Time: 451ms, Memory: 28MB, Status: Passed - CodeTree

+++ 240409/나무박멸/tree-kill-all.py
@@ -6,30 +6,33 @@ graph = [list(map(int, input().split())) for _ in range(N)]
 
 k_graph = [[0] * N for _ in range(N)]
 
-dx = [0,1,0,-1]
-dy = [1,0,-1,0]
-
 answer = 0
 
 def tree_grow():
+    dx = [0,1,0,-1]
+    dy = [1,0,-1,0]
     tmp = []
     for i in range(N):
         for j in range(N):
             if graph[i][j] > 0:
                 tmp.append((i,j))
     
-    q = deque()
-    for t_x, t_y in tmp:
-        cnt = 0
+    if len(tmp):
+        for t_x, t_y in tmp:
+            cnt = 0
 
-        for k in range(4):
-            nx, ny = t_x + dx[k], t_y + dy[k]
-            if 0 <= nx < N and 0 <= ny <N and graph[nx][ny] > 0:
-                cnt += 1
-        
-        graph[t_x][t_y] += cnt
+            for k in range(4):
+                nx, ny = t_x + dx[k], t_y + dy[k]
+                if 0 <= nx < N and 0 <= ny <N and graph[nx][ny] > 0:
+                    cnt += 1
+            
+            graph[t_x][t_y] += cnt
+    else:
+        return False
 
 def bunsik():
+    dx = [0,1,0,-1]
+    dy = [1,0,-1,0]
     tmp_graph = [[0] * N for _ in range(N)]
     tmp = []
     for i in range(N):
@@ -46,8 +49,8 @@ def bunsik():
                 cnt += 1
         
         if cnt > 0:
-            for k in range(4):
-                nx, ny = t_x + dx[k], t_y + dy[k]
+            for i in range(4):
+                nx, ny = t_x + dx[i], t_y + dy[i]
                 if 0 <= nx < N and 0 <= ny <N and graph[nx][ny] == 0 and k_graph[nx][ny] == 0:
                     tmp_graph[nx][ny] += (graph[t_x][t_y] // cnt)
                 
@@ -100,29 +103,30 @@ def kill(x,y):
 
     answer += graph[x][y]
     graph[x][y] = 0
-    k_graph[x][y] = C
+    k_graph[x][y] = C+1
 
     for i in range(4):
         for j in range(1,K+1):
             nx, ny = x + dx[i] * j, y + dy[i] * j
             if 0 <= nx < N and 0 <= ny < N:
                 if graph[nx][ny] == -1 or graph[nx][ny] == 0:
-                    k_graph[nx][ny] = C
+                    k_graph[nx][ny] = C+1
                     break
                 else:
                     answer += graph[nx][ny]
-                    k_graph[nx][ny] = C
+                    k_graph[nx][ny] = C+1
                     graph[nx][ny] = 0
-
 
 def remove_killer():
     for i in range(N):
         for j in range(N):
             if k_graph[i][j] > 0:
                 k_graph[i][j] -= 1
-
+                
 for _ in range(M):
-    tree_grow()
+    x = tree_grow()
+    if x == False:
+        continue
     bunsik()
     x,y = killer_select()
     kill(x,y)
